requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
distro==1.9.0
filelock==3.28.0
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.10.2
idna==3.11
jiter==0.14.0
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.4
openai==2.31.0
packaging==26.1
pydantic==2.13.0
pydantic_core==2.46.0
Pygments==2.20.0
python-dotenv==1.2.2
PyYAML==6.0.3
regex==2026.4.4
rich==15.0.0
safetensors==0.7.0
shellingham==1.5.4
sniffio==1.3.1
tokenizers==0.22.2
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing_extensions==4.15.0
typing-inspection==0.4.2
torch
accelerate
